Rename title-suggestion route file to titles, move api handler folder to title-suggestions and rename
