+++ src/routes/titles.js
@@ -1,9 +1,9 @@
 const {getSuggestionsWithCountedVotes} = require('../../database/suggestions');
-const {startNewVoteFromApi} = require('../webhook-handlers/vote');
+const {startNewVoteFromApi} = require('../../title-suggestions/api/vote');
 const {
     addNewSuggestionFromApi,
     removeSuggestionFromApi,
-} = require('../webhook-handlers/title-suggestions');
+} = require('../../title-suggestions/api/title-suggestions');
 const {adminPreHandler} = require('./loginUtilities');
 
 module.exports = (fastify, opts, done) => {
@@ -24,7 +24,6 @@ module.exports = (fastify, opts, done) => {
     // API: get suggestions
     fastify.get(
         '/api/titles/:epNum',
-        {preHandler: adminPreHandler},
         async (request, reply) => {
             const epNum = request.params.epNum;
             const countedSuggestions = await getSuggestionsWithCountedVotes({
@@ -34,16 +33,22 @@ module.exports = (fastify, opts, done) => {
         },
     );
 
-    // API: edit suggestions
+    // API: add suggestion
     fastify.post(
         '/api/titles/:messageId',
         {preHandler: adminPreHandler},
         (request, reply) => {
             const messageId = request.params.messageId;
-            addNewSuggestionFromApi(messageId);
-            reply.send(1);
+            addNewSuggestionFromApi(messageId)
+                .then(() => reply.send(1))
+                .catch((error) => {
+                    request.log.error({error, messageId});
+                    return reply.code(422); // unprocessable entity
+                });
         },
     );
+
+    // API: start vote
     fastify.post(
         '/api/vote',
         {preHandler: adminPreHandler},
@@ -52,6 +57,8 @@ module.exports = (fastify, opts, done) => {
             reply.send(1);
         },
     );
+
+    // API: delete suggestion
     fastify.delete(
         '/api/titles/:messageId',
         {preHandler: adminPreHandler},
